Add chrome and solr images with build system
We create a build system that automatically builds our images in
different versions so that we can easily use them in our CI and update
when needed. We create a workflow per image so that we only rebuild the
image that's actually been updated.

The images we add is a copy of `kingdutch/social-docker-chrome` as it's
currently being used as well as a new SOLR image which matches what we
use in our development docker-compose but is needed for the CI to be
able to use SOLR as we can't change the entrypoint from the GitHub
workflow configuration.